refactor(core): Relocate history model class
ref: #765 #759
